Modify areas with variable "amount" from type "double" to type "BigDecimal"
prints up to two decimal places with HALF_EVEN rounding.

+++ src/main/java/seedu/duke/record/Expense.java
@@ -1,12 +1,13 @@
 package seedu.duke.record;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Expense extends Record {
     private static final String TYPE_EXPENSE = "E";
     private static final String FILE_OUTPUT_STRING_FORMAT = "| %s | %s | %f | %s";
 
-    public Expense(double amount, LocalDate issuedDate, String description) {
+    public Expense(BigDecimal amount, LocalDate issuedDate, String description) {
         super(amount, issuedDate, description);
     }
